Created Map component
